Follow unfollow feature is implemented using ajax

+++ users/views.py
@@ -153,33 +153,44 @@ class HomeView(ListView):
 
 @login_required
 def other_user_profile(request, pk):
-    if (request.user.id == pk):
+
+    if request.user.id == pk:
         return HttpResponseRedirect('/profile')
 
     post = Post.objects.filter(author_id=pk)
     user = User.objects.get(id=pk)
 
     profile_other = Profile.objects.filter(id=pk).first()
-    profile_current = Profile.objects.filter(id=request.user.id).first()
+    profile_current = Profile.objects.filter(user=request.user).first()
 
     following_list = profile_current.following.all()
 
     following_list_of_other_user = profile_other.following.all()
-    follower_list_of_other_user = profile_other.following.all()
+    follower_list_of_other_user = profile_other.followers.all()
 
     button_status = 'none'
+    button_class = 'unfollowBtn'
+    button_id = 'unfollow-user'
+    button_text = 'Unfollow'
 
     if user.id not in [user for id, user in profile_current.following.values_list()]:
 
         button_status = 'not_following'
+        button_class = 'followBtn'
+        button_id = 'follow-user'
+        button_text = 'Follow'
 
-        if len(FollowRequest.objects.filter(from_user=request.user).
-                       filter(to_user=user)) == 1:
+        if len(FollowRequest.objects.filter(from_user=request.user).filter(to_user=user)) == 1:
             button_status = 'requested'
+            button_class = 'cancelRequestBtn'
+            button_id = 'cancel-request'
+            button_text = 'Cancel request'
 
     other_user_article_category = [name for id, name in Profile.objects.get(user=user).article_category.values_list()]
     current_user_article_category = [name for id, name in Profile.objects.get(user=request.user).article_category.values_list()]
     common_topics = [name for name in current_user_article_category if name in other_user_article_category]
+
+    current_user_following_list = [user for id, user in request.user.profile.following.values_list()]
 
     following_count = profile_other.following_count
     follower_count = profile_other.followers_count
@@ -191,32 +202,37 @@ def other_user_profile(request, pk):
         'user_id': user,
         'article_category': other_user_article_category,
         'button_status': button_status,
+        'button_class':button_class,
+        'button_id' : button_id,
+        'button_text':button_text,
         'following_count': following_count,
         'follower_count': follower_count,
         'common_topics':common_topics,
         'following_list': following_list_of_other_user,
         'follower_list': follower_list_of_other_user,
+        'following_list_of_current_user': current_user_following_list,
         'comment_form': comment_form
+
     }
 
     return render(request, 'users/otherUserProfile.html', context)
 
 
 """  This function gets called when the user follows some other user. Url path in Blogging/urls and the
- name of the url is "send_follow_request"   """
+ name of the url is "send_follow_request.js"   """
 
 
 @login_required
-def send_follow_request(request, pk):
+def send_follow_request(request):
     if request.user.is_authenticated:
-        user = get_object_or_404(User, id=pk)
+        userid = request.GET['userid']
+        user = get_object_or_404(User, id=userid)
         frequest, created = FollowRequest.objects.get_or_create(
             from_user=request.user,
             to_user=user
         )
 
-        return HttpResponseRedirect('/other-profile/{}/'.format(pk))
-
+        return render(request, 'users/userprofile.html')
 
 """  This function gets called when the user cancels the follow request sent to  some other user. 
 This wil  only be accessible after sending the follow request and before the acceptance of the request.
@@ -224,15 +240,17 @@ Url path in Blogging/urls and the name of the url is "cancel_follow_request"   "
 
 
 @login_required
-def cancel_follow_request(request, pk):
+def cancel_follow_request(request):
     if request.user.is_authenticated:
-        user = get_object_or_404(User, id=pk)
+        userid = request.GET['userid']
+        user = get_object_or_404(User, id=userid)
         frequest = FollowRequest.objects.filter(
             from_user=request.user,
             to_user=user
         ).first()
         frequest.delete()
-        return HttpResponseRedirect('/other-profile/{}/'.format(pk))
+        button_status = 'not_following'
+        return render(request, 'users/userprofile.html')
 
 
 """This view is called when the user who received the follow request deletes the request and this view 
@@ -241,7 +259,7 @@ url is "delete_follow_request" """
 
 
 @login_required
-def delete_follow_request(request, pk):
+def delete_follow_request(request):
     from_user = get_object_or_404(User, id=pk)
     frequest = FollowRequest.objects.filter(
         from_user=from_user,
@@ -259,23 +277,32 @@ only if the user follows that particular user. If this view is called then the u
 
 
 @login_required
-def unfollow_user(request, pk):
-    user = get_object_or_404(User, id=pk)
+def unfollow_user(request):
+    if request.user.is_authenticated:
+        userid = request.GET['userid']
+        user = get_object_or_404(User, id=int(userid))
 
-    unfollowed_user = Profile.objects.get(id=pk)
-    follower_list = [user for id, user in unfollowed_user.followers.values_list()]
-    follower_user_index = follower_list.index(request.user.id)
-    unfollowed_user.followers.all()[follower_user_index].delete()
-    unfollowed_user.followers_count -= 1
-    unfollowed_user.save()
+        unfollowed_user = Profile.objects.get(id=int(userid))
+        follower_list = [user for id, user in unfollowed_user.followers.values_list()]
+        follower_user_index = follower_list.index(request.user.id)
+        unfollowed_user.followers.all()[follower_user_index].delete()
+        unfollowed_user.followers_count -= 1
+        unfollowed_user.save()
 
-    follower = Profile.objects.get(user=request.user)
-    following_list = [user for id, user in follower.following.values_list()]
-    following_user_index = following_list.index(pk)
-    follower.following.all()[following_user_index].delete()
-    follower.following_count -= 1
-    follower.save()
-    return HttpResponseRedirect('/other-profile/{}/'.format(pk))
+        follower = Profile.objects.get(user=request.user)
+        following_list = [user for id, user in follower.following.values_list()]
+        following_user_index = following_list.index(int(userid))
+        follower.following.all()[following_user_index].delete()
+        follower.following_count -= 1
+        follower.save()
+
+        following_count = follower.following_count
+
+        data_dict = {'following_count':following_count}
+
+        return JsonResponse(data=data_dict, safe=False)
+
+    return render(request, 'users/userprofile.html')
 
 
 """This view gets called if the user who received the follow request accepts it. This is accessible 
@@ -286,25 +313,53 @@ who accepted the request will be added to the following list of the user who sen
 
 @login_required
 def accept_follow_request(request, pk):
-    from_user = get_object_or_404(User, id=pk)
+    if request.user.is_authenticated:
+        from_user = get_object_or_404(User, id=pk)
 
-    frequest = FollowRequest.objects.filter(
-        from_user=from_user, to_user=request.user).first()
+        frequest = FollowRequest.objects.filter(
+            from_user=from_user, to_user=request.user).first()
 
-    request_sender = from_user
-    requested_user = frequest.to_user
+        request_sender = from_user
+        requested_user = request.user
 
-    request_sender.profile.following.create(user=requested_user)
-    request_sender.profile.following_count += 1
-    request_sender.profile.save()
+        request_sender.profile.following.create(user=requested_user)
+        request_sender.profile.following_count += 1
+        request_sender.profile.save()
 
-    requested_user.profile.followers.create(user=request_sender)
-    requested_user.profile.followers_count += 1
-    requested_user.profile.save()
+        requested_user.profile.followers.create(user=request_sender)
+        requested_user.profile.followers_count += 1
+        requested_user.profile.save()
 
-    frequest.delete()
+        frequest.delete()
 
     return HttpResponseRedirect('/profile')
+
+
+@login_required
+def remove_from_followers_list(request):
+    if request.user.is_authenticated:
+        userid = request.GET['userid']
+        user_to_be_removed = get_object_or_404(User, id=int(userid))
+
+        follower_list = [user for id, user in request.user.profile.followers.values_list()]
+        user_to_be_removed_index = follower_list.index(user_to_be_removed.id)
+        request.user.profile.followers.all()[user_to_be_removed_index].delete()
+        request.user.profile.followers_count -= 1
+        request.user.profile.save()
+
+        following_list_of_user_to_be_removed = [user for id, user in user_to_be_removed.profile.following.values_list()]
+        removing_user_index = following_list_of_user_to_be_removed.index(request.user.id)
+        user_to_be_removed.profile.following.all()[removing_user_index].delete()
+        user_to_be_removed.profile.following_count -=1
+        user_to_be_removed.profile.save()
+
+        removing_user_followers_count = request.user.profile.followers_count
+        data_dict = {'followers_count':removing_user_followers_count}
+
+        return JsonResponse(data=data_dict, safe=False)
+
+    return render(request, 'users/userprofile.html')
+
 
 
 class SelectFavouriteArticleCategoryView(LoginRequiredMixin, View):
